Added error handling for updatePerms
We use updatePerms after logging in to update any and all commands with their required permissions.

I added error handling to this section, and awaited the result as seemingly this fixed the issues I was having. eslint claims await is useless in that context, but testing contradicted that, so I kept it.